% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPolymodalSpectrumGenerator.R
\name{runPSG}
\alias{runPSG}
\title{Launch Polymodal Spectrum Generator's Shiny App}
\usage{
runPSG()
}
\value{
No value, opens a Shiny page
}
\description{
This function launches the Shiny app for the Polymodal Spectrum Generator
program
}
\references{
Winston Chang, Joe Cheng, JJ Allaire, Carson Sievert, Barret Schloerke, Yihui Xie, Jeff Allen, Jonathan McPherson, Alan Dipert and Barbara Borges (2021). shiny: Web Application
Framework for R. R package version 1.7.1. https://CRAN.R-project.org/package=shiny
}
