% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePS.R
\name{generatePS}
\alias{generatePS}
\title{Generates the Polymodal Spectrum}
\usage{
generatePS(MSAfilepath, Consensusfilepath, N)
}
\arguments{
\item{MSAfilepath}{A string representation of the filepath for the .fasta
file containing the aligned protein sequences}

\item{Consensusfilepath}{A string representation of the filepath for the
.txt file containing the protein consensus sequence}

\item{N}{The number of key variants you want to be visualized}
}
\value{
Output is a spectrum plot of the variants, position based on their
relatedness, 0 being the variant most similar to the consensus, and 1
being the variant with the fewest consensus matches
}
\description{
With a filepath to the multiple sequence alignment, and consensus sequence
for a set of proteins, and a specified n, which is the number of desired
variants to be shown, the program returns of plot of the n most significant
varaints in the set.
}
\details{
The variant which matches the consensus sequence the best is position at the
origin (0), and the variant which is the worst match with the consensus is at 1.
The other n - 2 variants are positioned based off of the relative number of
matches they have with the best and worst matches. The variants are also assigned
a strength, which is their number of matches with the best and worst, which is
used to give priority to the varaints with the most matches.
}
\references{
Bodenhofer U, Bonatesta E, Horejs-Kainrath C, Hochreiter S (2015). “msa: an R package for multiple sequence alignment.” Bioinformatics, 31(24), 3997–3999. doi: 10.1093/bioinformatics/btv494.
Murrell, P. (2005). R Graphics. Chapman & Hall/CRC Press.
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2021). Biostrings: Efficient manipulation of biological strings. R package version 2.62.0, https://bioconductor.org/packages/Biostrings.
R Core Team (2021). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
Winston Chang, Joe Cheng, JJ Allaire, Carson Sievert, Barret Schloerke, Yihui Xie, Jeff Allen, Jonathan McPherson, Alan Dipert and Barbara Borges (2021). shiny: Web Application
Framework for R. R package version 1.7.1. https://CRAN.R-project.org/package=shiny
}
