% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPolymodalSpectrumGenerator.R
\name{runPSG}
\alias{runPSG}
\title{Launch Polymodal Spectrum Generator's Shiny App}
\usage{
runPSG()
}
\value{
No value, opens a Shiny page
}
\description{
This function launches the Shiny app for the Polymodal Spectrum Generator
program
}
\examples{
\dontrun{

FastRPA::runFastRPA()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
